requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
gunicorn==20.0.4
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
whitenoise==6.1.0